requirements.txt
amqp==2.4.2
appdirs==1.4.3
asgiref==3.2.3
asn1crypto==1.4.0
awscli==1.20.57
beautifulsoup4==4.8.2
billiard==3.6.0.0
bleach==3.1.3
bleach-whitelist==0.0.10
boto==2.49.0
boto3==1.18.57
botocore==1.21.57
celery==4.3.0
certifi==2021.10.8
cffi==1.13.2
chardet==3.0.4
charset-normalizer==2.0.9
colorama==0.4.3
cryptography==2.8
cvelib==0.7.0
click==8.0.4
dictdiffer==0.8.1
Django==2.2.26
django-appconf==1.0.3
django-celery-results==1.0.4
django-countries==5.3.3
django-environ==0.4.5
django-qr-code==1.1.0
django-ses==1.0.3
django-storages==1.7.1
django-widget-tweaks==1.4.3
djangorestframework==3.10.3
docutils==0.14
ecdsa==0.13.2
envs==1.3
fs==2.4.11
fs-s3fs==1.1.1
future==0.17.1
idna==3.3
jmespath==0.9.3
kombu==4.5.0
lxml==4.5.0
M2Crypto==0.36.0
Markdown==3.1
pinax-messages==2.0.2
pip-autoremove==0.9.1
psycopg2-binary==2.8.2
pyasn1==0.4.8
pycparser==2.19
pycryptodome==3.14.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-gnupg==0.4.5
python-jose==3.2.0
pytz==2019.3
PyYAML==5.4.1
qrcode==6.1
redis==3.2.1
requests==2.26.0
rsa==4.6
s3transfer==0.5.0
simplejson==3.17.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.26.7
vine==1.3.0
watchtower==1.0.6
webencodings==0.5.1
setuptools
